refactor(backend): make  review changes
