--- apps/iaimc-backend/src/app/common/entities/permission.entity.ts
@@ -1,13 +1,0 @@
-import { Column, Entity, PrimaryGeneratedColumn } from 'typeorm';
-
-@Entity('permissions')
-export class PermissionEntity {
-  @PrimaryGeneratedColumn('uuid')
-  public id!: string;
-
-  @Column('varchar', {
-    length: 100,
-    unique: true,
-  })
-  public name!: string;
-}
