+++ apps/iaimc-backend/src/app/common/types/general.types.ts
@@ -1,0 +1,25 @@
+import { TypeOrmModuleOptions } from '@nestjs/typeorm';
+
+export interface ResponseObject<T = undefined> {
+  message?: string;
+  data?: T;
+}
+
+export type DatabaseConfig = {
+  database: TypeOrmModuleOptions;
+};
+
+export type LOGGER_TYPE = 'log' | 'warn' | 'error' | 'debug';
+
+export enum ReviewStatus {
+  Unsubmitted = 'Unsubmitted',
+  Unassigned = 'Unassigned',
+  Assigned = 'Assigned',
+  Approved = 'Approved',
+  Rejected = 'Rejected',
+}
+
+export enum Role {
+  Admin = 'Admin',
+  Customer = 'Customer',
+}
