+++ apps/iaimc-backend/src/app/common/types/material.types.ts
@@ -1,31 +1,3 @@
-import { TypeOrmModuleOptions } from '@nestjs/typeorm';
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-import { Multer } from 'multer';
-
-export enum FileStoreDirectory {
-  MODELS = 'models',
-  CONTACT_FORMS = 'contact_forms',
-}
-
-export type DatabaseConfig = {
-  database: TypeOrmModuleOptions;
-};
-
-export type MulterFile = Express.Multer.File;
-
-export enum Accepted3DModelType {
-  STL = 'stl',
-  STP = 'stp',
-  STEP = 'step',
-}
-
-export interface ResponseObject<T = undefined> {
-  message?: string;
-  data?: T;
-}
-
-export type LOGGER_TYPE = 'log' | 'warn' | 'error' | 'debug';
-
 export type ResistanceLevel = 'x' | 'o' | '+' | '-';
 
 export interface ChemicalResistance {
@@ -79,16 +51,3 @@ export interface InjectionMoldingMaterial {
 
 export type InjectionMoldingMaterialPropertyValueType =
   InjectionMoldingMaterial[keyof InjectionMoldingMaterial];
-
-export enum ReviewStatus {
-  Unsubmitted = 'Unsubmitted',
-  Unassigned = 'Unassigned',
-  Assigned = 'Assigned',
-  Approved = 'Approved',
-  Rejected = 'Rejected',
-}
-
-export enum Role {
-  Admin = 'Admin',
-  Customer = 'Customer',
-}
