+++ apps/iaimc-backend/src/app/configurations/database.config.ts
@@ -1,12 +1,10 @@
 import { TypeOrmModuleOptions } from '@nestjs/typeorm';
-import { PermissionEntity } from '../common/entities/permission.entity';
 import { RoleEntity } from '../common/entities/role.entity';
 import { UserEntity } from '../common/entities/user.entity';
-import { DatabaseConfig } from '../common/types';
+import { DatabaseConfig } from '../common/types/general.types';
 import { FeedbackEntity } from '../modules/feedback/feedback.entity';
 import { ModelConfigurationEntity } from '../modules/model-configurations/entities/configuration.entity';
 import { FileEntity } from '../modules/model-configurations/entities/file.entity';
-import { PrincipalFormEntity } from '../modules/model-configurations/entities/principal-form.entity';
 
 export default (): DatabaseConfig => ({
   database: {
@@ -20,8 +18,6 @@ export default (): DatabaseConfig => ({
       FeedbackEntity,
       FileEntity,
       ModelConfigurationEntity,
-      PermissionEntity,
-      PrincipalFormEntity,
       RoleEntity,
       UserEntity,
     ],
