+++ apps/iaimc-backend/src/app/modules/model-configurations/entities/file.entity.ts
@@ -4,7 +4,7 @@ import {
   Entity,
   PrimaryGeneratedColumn,
 } from 'typeorm';
-import { Accepted3DModelType } from '../../../common/types';
+import { Accepted3DModelType } from '../../../common/types/file.types';
 
 @Entity('files')
 export class FileEntity {
