--- apps/iaimc-backend/src/app/modules/model-configurations/entities/principal-form.entity.ts
@@ -1,25 +1,0 @@
-import {
-  Column,
-  CreateDateColumn,
-  Entity,
-  PrimaryGeneratedColumn,
-  UpdateDateColumn,
-} from 'typeorm';
-
-@Entity('principal_forms')
-export class PrincipalFormEntity {
-  @PrimaryGeneratedColumn('uuid')
-  public id!: string;
-
-  @Column()
-  public name!: string;
-
-  @Column('int', { array: true })
-  public maxDimensions!: number[];
-
-  @CreateDateColumn()
-  public createdAt!: Date;
-
-  @UpdateDateColumn()
-  public updatedAt!: Date;
-}
